kagi: unify search output flags (--json/--text/--markdown); share _text helpers
Give `search` the same mutually-exclusive format group as `summarize`
(-j aliases --json), and extract the HTML/Markdown->text helpers into a
shared _text.py used by both verbs. Backward-compatible: the `search -j`
JSON keeps `raw_text` (alias of `markdown`) that released kagi-search emitted.
